execute package entry point in separate file
